+++ src/routing.ts
@@ -1,8 +1,8 @@
 import { Express } from "express";
-import { Logger } from "winston";
 import { CommonConfig } from "./config-type";
 import { DependsOnMethod } from "./depends-on-method";
 import { AbstractEndpoint } from "./endpoint";
+import { AbstractLogger } from "./logger";
 import { walkRouting } from "./routing-walker";
 import { ServeStatic } from "./serve-static";
 import { getStartupLogo } from "./startup-logo";
@@ -18,7 +18,7 @@ export const initRouting = ({
   routing,
 }: {
   app: Express;
-  logger: Logger;
+  logger: AbstractLogger;
   config: CommonConfig;
   routing: Routing;
 }) => {